wireshark: update API changes for wireshark-1.8 or later
Weirdly, the previous commit (6f92cd1) for parsing srp_addrs triggers
this API call and results to make wireshark coredumps.

The API changes had been made at:
  https://code.wireshark.org/review/gitweb?p=wireshark.git;a=commitdiff;h=a2414d8909088ddb40c907886e725993e6baecb5#patch77

The affected code is not included to the wireshark main repository
so no need to forward port.

This commit will be necessary if you want to use this plugin
with wireshark-1.8 or later, such as on CentOS 6/7.